move defense action to actor
